+++ packages/extension-ui/src/messaging.spec.ts
@@ -7,14 +7,8 @@ import '@polkadot/extension-mocks/chrome';
 
 import type * as _ from '@polkadot/dev-test/globals.d.ts';
 
-import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
-import enzyme from 'enzyme';
-
 import { wakeUpServiceWorkerWrapper } from '../../extension-base/src/utils/portUtils.js';
 import { exportAccount } from './messaging.js';
-
-// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment,@typescript-eslint/no-unsafe-call
-enzyme.configure({ adapter: new Adapter() });
 
 describe('messaging sends message to background via extension port for', () => {
   beforeEach(() => {